+++ NEWS.md
@@ -1,3 +1,7 @@
+# jbplot 1.2.1
+
+* Added `geom_lm()`: wrapper for `ggplot2::geom_smooth()` that sets `method` to `lm()` and `formula` to `y ~ x`
+
 # jbplot 1.2.0
 
 Added ggplot2 convenience functions: